fix: move popular section label below controls row to fix mobile stacking
