+++ about.html
@@ -16,7 +16,7 @@
 <link rel="stylesheet" href="css/shared/topbar.css?v=55e85c86">
 <link rel="stylesheet" href="css/shared/site.css?v=86052a10">
 <link rel="stylesheet" href="css/shared/cards.css?v=14c56151">
-<link rel="stylesheet" href="css/index/index.css?v=ff4b9357">
+<link rel="stylesheet" href="css/index/index.css?v=ca1427ec">
 </head>
 <body>
