+++ app.html
@@ -57,6 +57,6 @@
 <script src="js/lib/topbar.js?v=46cb7ea1"></script>
 <script src="js/lib/toast.js?v=7bb23014"></script>
 <script src="js/lib/gh-gist.js?v=346e3ea7"></script>
-<script type="module" src="js/app/main.js?v=9e0f8145"></script>
+<script type="module" src="js/app/main.js?v=08e6e4bb"></script>
 </body>
 </html>
