+++ index.html
@@ -16,7 +16,7 @@
 <link rel="stylesheet" href="css/shared/topbar.css?v=55e85c86">
 <link rel="stylesheet" href="css/shared/site.css?v=86052a10">
 <link rel="stylesheet" href="css/shared/cards.css?v=14c56151">
-<link rel="stylesheet" href="css/index/index.css?v=ff4b9357">
+<link rel="stylesheet" href="css/index/index.css?v=ca1427ec">
 </head>
 <body>
 
@@ -166,7 +166,6 @@
          the pipeline). Hidden until data loads so the section never shows empty. -->
     <section class="popular-games" id="popular-games" hidden>
       <div class="pg-head">
-        <div class="section-label" id="pg-section-label">Popular on Steam</div>
         <div class="pg-filter-wrap" id="pg-filter-wrap">
           <button class="pg-filter-toggle-btn" id="pg-filter-toggle" type="button" aria-expanded="false">
             Filters<span class="pg-filter-badge" id="pg-filter-badge" hidden></span>
@@ -197,6 +196,7 @@
           </div>
         </div>
       </div>
+      <div class="section-label pg-section-heading" id="pg-section-label">Popular on Steam</div>
       <p class="popular-games-sub" id="pg-sub">Steam's most-played games and how they run on Linux through Proton.</p>
       <div class="pg-list" id="pg-list"></div>
       <div class="pg-load-more-row" id="pg-load-more"></div>
